IISIM 2.2.5; IISE 1.0.5: EFM/FHR Variability & Accelerations IISIM - Implemented default DeviceEFM strip speed (x10) IILIB (Patient model) - Fetal HR baseline variability and wandering (based on variability) implemented - Fetal HR acceleration implemented using boundaried wandering   - Boundaries adjust based on state of current rhythm (newly implemented flag) IISE: Implemented obstetric parameter changes into IISE

+++ II Simulator/Windows/DeviceIABP.axaml.cs
@@ -144,7 +144,7 @@ namespace IISIM {
             // Random helium tank remaining amount... it's for show!
             this.FindControl<TextBlock> ("lblHelium").Text = String.Format ("{0}: {1:0}%",
                 Utility.WrapString (Instance.Language.Localize ("IABP:Helium")),
-                II.Math.RandomDouble (20, 80));
+                II.Math.RandomInt (20, 80));
 
             Grid displayGrid = this.FindControl<Grid> ("displayGrid");
